+++ odoo/local-src/enfinfidu_report/__openerp__.py
@@ -1,0 +1,23 @@
+# -*- coding: utf-8 -*-
+# © 2016 Julien Coux (Camptocamp)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+{
+    'name': 'Enfinfidu Report Customization',
+    'version': '9.0.1.0.1',
+    'author': 'Camptocamp',
+    'license': 'AGPL-3',
+    'category': 'Specific',
+    'website': 'http://www.camptocamp.com',
+    'images': [],
+    'depends': [
+        'report',
+        'stock',
+    ],
+    'data': [
+        # Reports
+        'report/layout.xml',
+    ],
+    'test': [],
+    'installable': True,
+    'auto_install': False,
+}