+++ odoo/songs/install/accounting.py
@@ -27,41 +27,38 @@ def no_coa_instance_unlock(ctx):
 
 
 @anthem.log
-def configure_missing_chart_of_account(ctx):
+def configure_missing_chart_of_account(ctx, coa_dict=False):
     """Configure Missing COA for companies"""
 
-    coa_dict = {
-        'base.main_company': {
-            'chart_template_id': 'enfinfidu_account.enfinfidu_chart_template',
-            'template_transfer_account_id':
-                'enfinfidu_account.transfer_account_id',
-            # 'sale_tax_id': 'l10n_ch.1_vat_80',
-            # 'purchase_tax_id': 'l10n_ch.1_vat_80_purchase',
-        },
-        '__setup__.rlbatiment': {
-            'chart_template_id': 'enfinfidu_account.enfinfidu_chart_template',
-            'template_transfer_account_id':
-                'enfinfidu_account.transfer_account_id',
-            # 'sale_tax_id': 'l10n_ch.1_vat_80',
-            # 'purchase_tax_id': 'l10n_ch.1_vat_80_purchase',
-        },
-    }
+    if not coa_dict:
+        coa_dict = {
+            'base.main_company': {
+                'chart_template_id':
+                    'enfinfidu_account.enfinfidu_chart_template',
+                'template_transfer_account_id':
+                    'enfinfidu_account.transfer_account_id',
+                # 'sale_tax_id': 'l10n_ch.1_vat_80',
+                # 'purchase_tax_id': 'l10n_ch.1_vat_80_purchase',
+                'sale_tax_id': 'enfinfidu_account.tax_tmp_vat_80',
+                'purchase_tax_id': 'enfinfidu_account.tax_tmp_vat_80_purchase',
+            },
+        }
     for company_xml_id, values in coa_dict.iteritems():
         company = ctx.env.ref(company_xml_id)
         coa = ctx.env.ref(values['chart_template_id'])
         template_transfer_account = ctx.env.ref(
             values['template_transfer_account_id']
         )
-        # sale_tax = ctx.env.ref(values['sale_tax_id'])
-        # purchase_tax = ctx.env.ref(values['purchase_tax_id'])
+        sale_tax = ctx.env.ref(values['sale_tax_id'])
+        purchase_tax = ctx.env.ref(values['purchase_tax_id'])
         if not company.chart_template_id:
             wizard = ctx.env['wizard.multi.charts.accounts'].create({
                 'company_id': company.id,
                 'chart_template_id': coa.id,
                 'transfer_account_id': template_transfer_account.id,
                 'code_digits': coa.code_digits,
-                # 'sale_tax_id': sale_tax.id,
-                # 'purchase_tax_id': purchase_tax.id,
+                'sale_tax_id': sale_tax.id,
+                'purchase_tax_id': purchase_tax.id,
                 'complete_tax_set': coa.complete_tax_set,
                 'currency_id': ctx.env.ref('base.CHF').id,
                 'bank_account_code_prefix': coa.bank_account_code_prefix,
@@ -71,6 +68,14 @@ def configure_missing_chart_of_account(ctx):
 
 
 @anthem.log
+def delete_l10n_ch_tags(ctx):
+    tags = ctx.env['account.account.tag'].search([
+        ('name', '=like', 'DEPRECATED%')])
+    tags.unlink()
+
+
+@anthem.log
 def main(ctx):
     """ Configuring accounting """
     configure_missing_chart_of_account(ctx)
+    delete_l10n_ch_tags(ctx)